TD-1379: refactor tools into separate modules
