+++ apps/chat-bot/src/app/api/chat/tools/retrieve-entire-file-tool.ts
@@ -1,0 +1,115 @@
+import { RETRIEVE_ENTIRE_FILE_CHARACTER_LIMIT } from '@/configuration-text-inputs/const';
+import { dbGetExtractedFileContent } from '@shared/db/functions/files';
+import type { FileModelAndContent } from '@shared/db/schema';
+import type { BuildToolsContext, ToolDefinition, ToolRegistration } from './types';
+
+type RetrieveEntireFileToolResponse = {
+  fileName: string | null;
+  content: string | null;
+  truncated: boolean;
+  characterCount: number;
+  maxCharacters: number;
+  error: string | null;
+};
+
+function truncateToCharacterLimit(text: string, maxCharacters: number) {
+  if (text.length <= maxCharacters) {
+    return text;
+  }
+
+  return text.slice(0, maxCharacters);
+}
+
+function formatEntireFileForTool(file: FileModelAndContent) {
+  const content = file.content?.trim() ?? '';
+  const characterCount = content.length;
+  const truncatedContent = truncateToCharacterLimit(content, RETRIEVE_ENTIRE_FILE_CHARACTER_LIMIT);
+  const truncated = truncatedContent.length !== content.length;
+
+  const response: RetrieveEntireFileToolResponse = {
+    fileName: file.name ?? null,
+    content: content.length > 0 ? truncatedContent : null,
+    truncated,
+    characterCount,
+    maxCharacters: RETRIEVE_ENTIRE_FILE_CHARACTER_LIMIT,
+    error: null,
+  };
+
+  if (!content) {
+    response.error = 'No usable content found.';
+  } else if (truncated) {
+    response.error = 'File content was truncated to fit the character limit.';
+  }
+
+  return JSON.stringify(response);
+}
+
+type BuildRetrieveEntireFileToolParams = Pick<BuildToolsContext, 'relatedFileEntities'>;
+
+export function buildRetrieveEntireFileTool({
+  relatedFileEntities,
+}: BuildRetrieveEntireFileToolParams): ToolRegistration | null {
+  if (relatedFileEntities.length === 0) {
+    return null;
+  }
+
+  const attachedFileDescriptions = relatedFileEntities.map(
+    (file) => `${file.name} (${file.size} bytes)`,
+  );
+
+  const definition: ToolDefinition = {
+    name: 'retrieve_entire_file',
+    description: `Retrieve the full content of one attached file by name. Available files right now: ${attachedFileDescriptions.join(', ') || 'none'}. Use this tool when you need the full text of a specific attached file instead of only relevant excerpts. The returned content is capped at ${RETRIEVE_ENTIRE_FILE_CHARACTER_LIMIT} characters.`,
+    parameters: {
+      type: 'object',
+      properties: {
+        fileName: {
+          type: 'string',
+          description: 'The exact name of the attached file to retrieve.',
+        },
+      },
+      required: ['fileName'],
+      additionalProperties: false,
+    },
+  };
+
+  const handler = async (args: Record<string, unknown>) => {
+    const fileName = typeof args.fileName === 'string' ? args.fileName.trim() : '';
+
+    if (fileName.length === 0) {
+      const response: RetrieveEntireFileToolResponse = {
+        fileName: null,
+        content: null,
+        truncated: false,
+        characterCount: 0,
+        maxCharacters: RETRIEVE_ENTIRE_FILE_CHARACTER_LIMIT,
+        error: 'Missing file name.',
+      };
+
+      return JSON.stringify(response);
+    }
+
+    const matchedFile = relatedFileEntities.find((file) => file.name === fileName);
+
+    if (matchedFile === undefined) {
+      const response: RetrieveEntireFileToolResponse = {
+        fileName,
+        content: null,
+        truncated: false,
+        characterCount: 0,
+        maxCharacters: RETRIEVE_ENTIRE_FILE_CHARACTER_LIMIT,
+        error: 'File not found.',
+      };
+
+      return JSON.stringify(response);
+    }
+
+    if (matchedFile.content === undefined) {
+      matchedFile.content = await dbGetExtractedFileContent(matchedFile.id);
+    }
+
+    return formatEntireFileForTool(matchedFile);
+  };
+
+  return { definition, handler };
+}
