+++ apps/chat-bot/src/app/api/chat/tools/types.ts
@@ -1,0 +1,23 @@
+import { type ToolDefinition, type ToolRegistry } from '@ais-chat/ai-core';
+import type { UserAndContext } from '@/auth/types';
+import type { FileModelAndContent } from '@shared/db/schema';
+import type { WebSearchResult } from '@shared/db/schema';
+
+export type { ToolDefinition, ToolRegistry };
+
+export type ToolRegistration = {
+  definition: ToolDefinition;
+  handler: (args: Record<string, unknown>) => Promise<string>;
+};
+
+export type BuildToolsContext = {
+  user: UserAndContext;
+  characterId?: string;
+  learningScenarioId?: string;
+  assistantId?: string;
+  conversationId: string;
+  relatedFileEntities: FileModelAndContent[];
+  sourceUrls: string[];
+  attachedLinks: string[];
+  onWebSearchResults?: (results: WebSearchResult[]) => void;
+};
